copy_engine/requirements.txt
Flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
Flask-Login>=0.6.0
Werkzeug>=2.3.0
openai>=1.30.0
python-dotenv>=1.0.0
Pillow>=10.0.0
requests>=2.31.0
